+++ app/models/user.rb
@@ -14,7 +14,37 @@ class User < ApplicationRecord
   # Validations
   validates :full_name, presence: true
 
+  # Uploaders
+  mount_uploader :consent_signature, SignatureUploader
+
   # Methods
+
+  def consent!(data_uri)
+    save_signature!(data_uri)
+    update consented_at: Time.zone.now
+  end
+
+  def consented?
+    !consented_at.nil?
+  end
+
+  def save_signature!(data_uri)
+    file = Tempfile.new("consent_signature.png")
+    begin
+      encoded_image = data_uri.split(",")[1]
+      decoded_image = Base64.decode64(encoded_image)
+      # File.open(file.path, "wb") ???
+      File.open(file, "wb") { |f| f.write(decoded_image) }
+      file.define_singleton_method(:original_filename) do
+        "consent_signature.png"
+      end
+      self.consent_signature = file
+      save
+    ensure
+      file.close
+      file.unlink # deletes the temp file
+    end
+  end
 
   def send_welcome_email_in_background(pw)
     fork_process(:send_welcome_email, pw)