+++ config/initializers/app_config.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# For saving signatures.
+require "base64"