+++ db/migrate/20170714143900_add_consented_at_to_users.rb
@@ -1,0 +1,5 @@
+class AddConsentedAtToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :consented_at, :datetime
+  end
+end