+++ db/migrate/20170714163736_add_consent_signature_to_users.rb
@@ -1,0 +1,5 @@
+class AddConsentSignatureToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :consent_signature, :text
+  end
+end